\name{Histogram}
\alias{Histogram}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Calculate Histogram
}            
\description{
Calculate Histogram of a numerical vector xh
}
\usage{
histogram(hx)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector of numeric values
}
}
\details{
Calculate Histogram                                                         l
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
Input integer vector
}
\author{
Md. Mahabubur Rahman
}
